+++ src/components/FeedList/FeedReaction.jsx
@@ -1,3 +1,4 @@
+import { postReaction } from '../../api/post';
 import styles from '../FeedList/FeedReaction.module.css';
 import { useState } from 'react';
 
@@ -6,22 +7,26 @@ export function FeedReaction({ like, dislike }) {
   const [countDislike, setCountDislike] = useState(dislike);
   const [activeButton, setActiveButton] = useState(null);
 
-  // 낙관적 업데이트 유지 & 서버에 해당하는 question ID 에 like, dislike 정보 업데이트하기 (PATCH)
+  // 낙관적 업데이트 유지 & 서버에 해당하는 question ID 에 like, dislike 추가 (POST)
   const onClickLike = () => {
     if (activeButton === 'like') {
       setCountLike(prevCount => prevCount + 1);
+      postReaction('like');
     } else {
       setActiveButton('like');
       setCountLike(prevCount => prevCount + 1);
+      postReaction('like');
     }
   };
 
   const onClickDislike = () => {
     if (activeButton === 'dislike') {
       setCountDislike(prevCount => prevCount + 1);
+      postReaction('dislike');
     } else {
       setActiveButton('dislike');
       setCountDislike(prevCount => prevCount + 1);
+      postReaction('dislike');
     }
   };
